requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
chromadb>=0.4.15
google-generativeai>=0.3.0
pypdf>=3.17.0
python-dotenv>=1.0.0
numpy>=1.21.0,<1.25.0
pandas>=1.3.0
tiktoken>=0.5.0
